requirements.txt
python-dotenv
langchain
faiss-cpu
langchain-community
langchain-google-genai
langchain-core 
langchain-text-splitters